refactor(services): integrates external service config into view
